Edited the tests and amended the database schema

+++ gateway/database.py
@@ -1,23 +1,25 @@
 import uuid
-
+import json
 import psycopg2
 from psycopg2.extras import register_uuid
 
 class PostgresDB:
     """This class is used to store the names of iot devices and their corresponding topics in a database.
     """
-    def __init__(self,
-                 host="161.35.205.102", 
-                 user="karelia", 
-                 password="postgres", 
-                 database="iot_gateway"):
+    def __init__(self):
         try:
+            with open("config.json") as f:
+                config = json.load(f) 
+                self.connection = psycopg2.connect(host=config["postgres_setup"]["host"],
+                                                   user=config["postgres_setup"]["user"], 
+                                                   password=config["postgres_setup"]["password"], 
+                                                   database=config["postgres_setup"]["database"])
             register_uuid()
-            self.connection = psycopg2.connect(host=host, user=user, password=password, database=database)
             self.cursor = self.connection.cursor()
             self.cursor.execute("""CREATE TABLE IF NOT EXISTS devices (
                                 id UUID PRIMARY KEY,
-                                device_id VARCHAR(255) UNIQUE NOT NULL,
+                                device_id VARCHAR(255) NOT NULL,
+                                jsonpath VARCHAR(255) NOT NULL,
                                 topic VARCHAR(255) NOT NULL
 )
 """)
@@ -27,7 +29,7 @@ class PostgresDB:
             exit(1)
 
     
-    def add_device(self, device_id, topic):
+    def add_device(self, device_id, jsonpath, topic):
         """Add a device to the database.
 
         Args:
@@ -36,16 +38,16 @@ class PostgresDB:
         """
         try:
             device_uuid = uuid.uuid4()
-            self.cursor.execute("INSERT INTO devices (id, device_id, topic) VALUES (%s, %s, %s)", (device_uuid, device_id, topic))
+            self.cursor.execute("""INSERT INTO devices (id, device_id, jsonpath, topic) VALUES (%s, %s, %s, %s)""", (device_uuid, device_id, jsonpath, topic))
 
             self.connection.commit()
         
         except psycopg2.IntegrityError:
             print(f"Device {device_id} already exists!")
             self.connection.rollback()
-        
-    def get_topic(self, device_id):
-        """Get the topic of a device.
+            
+    def get_jsonpath(self, device_id, topic):
+        """Get the jsonpath of a device.
         
         Args:
             topic (str): The topic of the device.
@@ -54,12 +56,27 @@ class PostgresDB:
             str: The id of the device.
         """
         try:
+            self.cursor.execute("""SELECT jsonpath FROM devices WHERE device_id=%s AND topic=%s""", (device_id, topic))
+            return self.cursor.fetchone()
+        except TypeError:
+            print(f"Device {device_id} does not exist!")
+        
+    def get_topic(self, device_id):
+        """Get the topic of a device.
+        
+        Args:
+            device_id (str): The id of the device.
+        
+        Returns:
+            str: The topic of the device.
+        """
+        try:
             self.cursor.execute("""SELECT topic FROM devices WHERE device_id=%s""", (device_id,))
             return self.cursor.fetchone()
         except TypeError:
             print(f"Device {device_id} does not exist!")
     
-    def get_device_id(self, topic):
+    def get_device_id(self, jsonpath, topic):
         """Get the id of a device.
         
         Args:
@@ -69,25 +86,25 @@ class PostgresDB:
             str: The id of the device.
         """
         try:
-            self.cursor.execute("""SELECT device_id FROM devices WHERE topic=%s""", (topic,))
+            self.cursor.execute("""SELECT device_id FROM devices WHERE jsonpath=%s AND topic=%s""", (jsonpath, topic))
             return self.cursor.fetchone()
         except TypeError:
-            print(f"Topic {topic} does not exist!")
+            print("Device does not exist!")
             
     def get_all_devices(self):
         """Get all devices from the database."""
         self.cursor.execute("""SELECT * FROM devices""")
         return self.cursor.fetchall()
     
-    def delete_device(self, device_id):
+    def delete_device(self, device_id, topic):
         """Delete a device from the database.
 
         Args:
             device_id (str): The id of the device.
+            topic (str): The topic of the device.
         """
-        
         try:
-            self.cursor.execute("""DELETE FROM devices WHERE device_id=%s""", (device_id,))
+            self.cursor.execute("""DELETE FROM devices WHERE device_id=%s AND topic=%s""", (device_id, topic))
             self.connection.commit()
         except psycopg2.IntegrityError:
             print(f"Device {device_id} does not exist!")
@@ -102,7 +119,7 @@ class PostgresDB:
             print("No devices to delete!")
             self.connection.rollback()
             
-    def update_device(self, device_id, topic):
+    def update_device(self, jsonpath, topic, device_id):
         """Update a device in the database.
 
         Args:
@@ -110,7 +127,7 @@ class PostgresDB:
             topic (str): The topic of the device.
         """
         try:
-            self.cursor.execute("""UPDATE devices SET topic=%s WHERE device_id=%s""", (topic, device_id))
+            self.cursor.execute("""UPDATE devices SET jsonpath=%s, topic=%s WHERE device_id=%s""", (jsonpath, topic, device_id))
             self.connection.commit()
         except psycopg2.IntegrityError:
             print(f"Device {device_id} does not exist!")
@@ -119,6 +136,10 @@ class PostgresDB:
     def nuke_table(self):
         """Delete the devices table."""
         try:
+            print("This will delete the devices table! Are you sure? (y/N)")
+            if input() != "y":
+                print("Aborting...")
+                return
             self.cursor.execute("""DROP TABLE devices""")
             self.connection.commit()
         except psycopg2.IntegrityError:
@@ -136,8 +157,5 @@ class PostgresDB:
         self.close()
 
 if __name__ == "__main__":
-    database = PostgresDB(host="161.35.205.102", user="karelia", password="postgres", database="iot_gateway")
-    database.add_device("sensor:001", "sensor___001")
-    database.add_device("sensor:002", "sensor___002")
-    print(database.get_all_devices())
-    database.close()
+    database = PostgresDB()
+    database.nuke_table()
